task2/requirements.txt
pinecone-client
torch
transformers
python-dotenv
pymupdf
